Fixes; on insert, more rigorously add the document to the in-memeory cache after the insert transaction completes.

+++ client-collection.coffee
@@ -23,7 +23,7 @@ if db.version is ''
 collections = {}
 
 Meteor.LocalMsg.listen
-  'Method.ClientCollection.single': (collection_name, doc_id) ->
+  'Meteor.ClientCollection.single': (collection_name, doc_id) ->
     collections[collection_name]?._reload_single(doc_id)
 
 Meteor.ClientSQLCollection = (name, cb) ->
@@ -72,7 +72,7 @@ _.extend Meteor.ClientSQLCollection.prototype,
           [@_name, doc_id],
           ((tx, result) =>
             if result.rows.length is 1
-              doc = result.rows.item(0)
+              doc = JSON.parse(result.rows.item(0).document)
           )
         )
       ),
@@ -87,13 +87,18 @@ _.extend Meteor.ClientSQLCollection.prototype,
     if doc._id?
       throw new Error 'inserted doc should not yet have an _id attribute'
     doc._id = LocalCollection.uuid()
-    @_localCollection.insert(doc)
     db.transaction(
       ((tx) =>
         tx.executeSql 'INSERT INTO documents (id, collection, document) VALUES (?, ?, ?)',
           [doc._id, @_name, JSON.stringify(doc)]
       ),
-      ((error) => console.log error)
+      ((error) =>
+        # might be out of space?
+        console.log 'insert transaction error', error
+      ),
+      (=>
+        @_localCollection.insert(doc)
+      )
     )
     doc._id
 
@@ -122,7 +127,6 @@ _.extend Meteor.ClientSQLCollection.prototype,
       ),
       ((error) => console.log 'modify transaction error', error),
       (=>
-        console.log 'modify transaction successful'
         if doc?
           @_localCollection.update doc._id, doc
           Meteor.LocalMsg.send 'Meteor.ClientCollection.single', @_name, doc._id
@@ -132,6 +136,9 @@ _.extend Meteor.ClientSQLCollection.prototype,
     )
 
 Meteor.ClientSQLCollection.erase = ->
+  # TODO not good if other windows already have a collection open
+  unless _.isEmpty(collections)
+    throw new Error("call erase() before opening any collections")
   db.transaction(
     ((tx) -> tx.executeSql('DELETE FROM documents')),
     ((error) -> console.log 'erase transaction error', error),
